Fix/보관함 다중삭제 에러 대응

+++ app/src/main/java/com/runnect/runnect/data/model/RequestPutMyDrawDto.kt
@@ -7,5 +7,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class RequestPutMyDrawDto(
     @SerialName("courseIdList")
-    val courseIdList: List<Long>
+    val courseIdList: List<Int>
 )
